trts: fix an overflow

+++ sdk/trts/ema_init.cpp
@@ -64,7 +64,7 @@ static int build_rts_context_nodes(layout_entry_t *entry, uint64_t offset)
     assert(IS_PAGE_ALIGNED(rva));
 
     size_t addr = (size_t)get_enclave_base() + rva;
-    size_t size = entry->page_count << SE_PAGE_SHIFT;
+    size_t size = ((size_t)entry->page_count) << SE_PAGE_SHIFT;
     size_t enclave_end = (size_t)get_enclave_base() + get_enclave_size();
 
     // entry is guard page or has EREMOVE, build a reserved ema
@@ -113,7 +113,7 @@ static int build_rts_context_nodes(layout_entry_t *entry, uint64_t offset)
         }
 
         int ret = mm_alloc((void*)addr,
-                         entry->page_count << SE_PAGE_SHIFT,
+                         ((size_t)entry->page_count) << SE_PAGE_SHIFT,
                          SGX_EMA_COMMIT_ON_DEMAND | commit_direction
                          | SGX_EMA_SYSTEM | SGX_EMA_FIXED | type,
                          NULL, NULL, NULL);
